backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
tensorflow-cpu==2.16.1
librosa==0.10.1
opencv-python-headless==4.9.0.80
numpy==1.24.3
pandas==2.2.0
pydub==0.25.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0